+++ uwadv/source/conv/codevm.hpp
@@ -97,7 +97,7 @@ public:
    ua_conv_globals(){}
 
    //! returns a list of globals for a given conv slot
-   std::vector<Uint8> &get_globals(Uint16 conv)
+   std::vector<Uint16> &get_globals(Uint16 conv)
    {
       if (conv>allglobals.size()) throw ua_ex_globals_access;
       return allglobals[conv];
@@ -110,7 +110,7 @@ public:
 
 protected:
    //! list with all globals from all conversations
-   std::vector< std::vector<Uint8> > allglobals;
+   std::vector< std::vector<Uint16> > allglobals;
 };
